grok_app/requirements.txt
fastapi 
uvicorn 
gradio 
openai 
python-dotenv 

